+++ src/main/java/com/gdg/group13/task/TodoListRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.time.LocalDate;
 import java.util.List;
 
-public interface TodoListRepository extends JpaRepository<TodoListEntity, Long> {
+public interface TodoListRepository extends JpaRepository<TodoListEntity, Integer> {
   TodoListEntity findByUidAndDate(Integer uid, LocalDate date);
 }
